Can do 1 and 2 length names,

+++ index.js
@@ -24,7 +24,7 @@ bot.on('message', (message, guild) => {
     const args = message.content.slice(prefix.length).split(' ');
     const command = args.shift().toLowerCase();
 
-    if (command == 'crole' ){
+    if (command == 'crole' && !args[1]){
         
         message.guild.roles.create(
         {
@@ -36,7 +36,20 @@ bot.on('message', (message, guild) => {
         })
             .then(console.log)
             .catch(console.error);
-            return message.reply ('You have made a new role.')
+            return message.reply ('You have made a new role.')}
+    if (command == 'crole'){
+        
+        message.guild.roles.create(
+                {
+            data: {
+                name : args[0] + ' ' + args[1],
+                color : Math.floor(Math.random() * 16777214) + 1,        
+                },
+            reason: 'We are Lazy.',       
+                })
+                    .then(console.log)
+                    .catch(console.error);
+                    return message.reply ('You have made a new role.')
 }});
 //list roles
 bot.on('message', (message, guild) => {
@@ -63,16 +76,28 @@ bot.on('message', (message, guild) => {
         const args = message.content.slice(prefix.length).split(' ');
         const command = args.shift().toLowerCase();
     
-    if (command == "arole"){
+    if (command == "arole" && !args[2])  { 
+        //if (args.length == 1){
         var user = message.mentions.members.first();
         var role= message.member.guild.roles.cache.find(role => role.name === args[0]);
         if(!role) return message.reply(args[0] + " aint a roll");
         if(!user) return message.reply ("Aint nobody named that.");
          
         else user.roles.add(role);
-          return message.reply ("Role Be There.")
-        }
-});
+          return message.reply ("Role Be There.")}
+
+    if (command == "arole")  { 
+        
+        var user = message.mentions.members.first();
+        var role= message.member.guild.roles.cache.find(role => role.name === args[0] + ' ' + args[1]);
+        if(!role) return message.reply(args[0] +' '+ args[1]+ " aint a roll");
+        if(!user) return message.reply ("Aint nobody named that.");
+             
+        else user.roles.add(role);
+        return message.reply ("Role Be There.")
+    
+    
+        }});
 //remove roles form user
 bot.on('message', (message, guild, client, ) => {
     if (!message.content.startsWith(prefix) || message.author.bot || !message.member.hasPermission('MANAGE_ROLES')) return;
@@ -80,14 +105,24 @@ bot.on('message', (message, guild, client, ) => {
         const args = message.content.slice(prefix.length).split(' ');
         const command = args.shift().toLowerCase();
 
-    if (command == "rrole"){
+    if (command == "rrole" && !args[2]){
         var user = message.mentions.members.first();
         var role= message.member.guild.roles.cache.find(role => role.name === args[0]);
         if(!role) return message.reply(args[0] + " aint a roll");
         if(!user) return message.reply ("Aint nobody named that.");
          
         else user.roles.remove(role);
-          return message.reply ("Role Be Gone.")
+          return message.reply ("Role Be Gone.")}
+
+          if (command == "rrole"){
+            var user = message.mentions.members.first();
+            var role= message.member.guild.roles.cache.find(role => role.name === args[0] + ' ' + args[1]);
+            if(!role) return message.reply(args[0] +' '+ args[1] + " aint a roll");
+            if(!user) return message.reply ("Aint nobody named that.");
+             
+            else user.roles.remove(role);
+              return message.reply ("Role Be Gone.")
+           
        
 }});
 //delete roles
@@ -97,12 +132,19 @@ bot.on('message', (message, guild, client, ) => {
         const args = message.content.slice(prefix.length).split(' ');
         const command = args.shift().toLowerCase();
 
-        if (command == "drole") {
+        if (command == "drole" && !args[1]) {
             var role= message.member.guild.roles.cache.find(role => role.name === args[0]);
             if(!role) return message.reply(args[0] + " aint a roll");
 
             else role.delete();
-            return message.reply("deleted " + args[0])
+            return message.reply("deleted " + args[0])}
+
+            if (command == "drole" ) {
+                var role= message.member.guild.roles.cache.find(role => role.name === args[0] + ' ' + args[1]);
+                if(!role) return message.reply(args[0] + ' ' + args[1] + " aint a roll");
+    
+                else role.delete();
+                return message.reply("deleted " + args[0] + ' ' + args[1])
 
         }
 
@@ -116,12 +158,19 @@ bot.on('message', (message, guild, client, ) => {
         const args = message.content.slice(prefix.length).split(' ');
         const command = args.shift().toLowerCase();
 
-    if (command == "parole"){
+    if (command == "parole" && !args[1]){
         var role= message.member.guild.roles.cache.find(role => role.name === args[0]);
         if(!role) return message.reply(args[0] + " aint a roll");
          
         else message.member.roles.add(role);
-          return message.reply ("Role be on you.")
+          return message.reply ("Role be on you.")}
+
+          if (command == "parole"){
+            var role= message.member.guild.roles.cache.find(role => role.name === args[0] + ' ' + args[1]);
+            if(!role) return message.reply(args[0] + ' ' + args[1] + " aint a roll");
+             
+            else message.member.roles.add(role);
+              return message.reply ("Role be on you.")
        
 }});
 //remove roles from self
@@ -131,11 +180,18 @@ bot.on('message', (message, guild, client, ) => {
         const args = message.content.slice(prefix.length).split(' ');
         const command = args.shift().toLowerCase();
 
-    if (command == "prrole"){
+    if (command == "prrole" && !args[1]){
         var role= message.member.guild.roles.cache.find(role => role.name === args[0]);
         if(!role) return message.reply(args[0] + " aint a roll");
         
         else message.member.roles.remove(role);
-          return message.reply ("Role be gone from you.")
+          return message.reply ("Role be gone from you.")}
+
+          if (command == "prrole"){
+            var role= message.member.guild.roles.cache.find(role => role.name === args[0] + ' ' + args[1]);
+            if(!role) return message.reply(args[0] + ' ' + args[1] + " aint a roll");
+            
+            else message.member.roles.remove(role);
+              return message.reply ("Role be gone from you.")
        
 }});
